# counter_app/openapi/openapi.json
{
    "components": {
        "schemas": {
            "CounterIncrement": {
                "description": "Schema for an increment request.",
                "properties": {
                    "amount": {
                        "title": "Amount",
                        "type": "integer"
                    }
                },
                "required": [
                    "amount"
                ],
                "title": "CounterIncrement",
                "type": "object"
            },
            "CounterValue": {
                "description": "Schema for representing the counter value.",
                "properties": {
                    "value": {
                        "title": "Value",
                        "type": "integer"
                    }
                },
                "required": [
                    "value"
                ],
                "title": "CounterValue",
                "type": "object"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "title": "Detail",
                        "type": "array"
                    }
                },
                "title": "HTTPValidationError",
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "title": "Location",
                        "type": "array"
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                },
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError",
                "type": "object"
            }
        }
    },
    "info": {
        "contact": {
            "email": "ferenc.toeroek.bp@gmail.com",
            "name": "Ferenc T\u00f6r\u00f6k"
        },
        "description": "A simple API to manage a single integer counter",
        "title": "Counter Service",
        "version": "1.0.0"
    },
    "openapi": "3.1.0",
    "paths": {
        "/counter": {
            "get": {
                "description": "Retrieve the current value of the counter. If no counter exists yet, one will be created automatically.",
                "operationId": "read_counter_counter_get",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CounterValue"
                                }
                            }
                        },
                        "description": "Counter value retrieved successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "summary": "Get current counter value",
                "tags": [
                    "counter",
                    "counter"
                ]
            }
        },
        "/counter/increment": {
            "post": {
                "description": "Increase the counter by a specified positive amount. The amount must be greater than zero.",
                "operationId": "increment_counter_increment_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CounterIncrement"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        },
                        "description": "Counter incremented successfully"
                    },
                    "400": {
                        "description": "Invalid increment amount; must be positive"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "summary": "Increment counter",
                "tags": [
                    "counter",
                    "counter"
                ]
            }
        },
        "/counter/reset": {
            "post": {
                "description": "Reset the counter to zero.",
                "operationId": "reset_counter_reset_post",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        },
                        "description": "Counter reset successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "summary": "Reset counter",
                "tags": [
                    "counter",
                    "counter"
                ]
            }
        }
    }
}